Use long even after %, but still 13/30 AC

+++ ADT_20240509_2_E/c.cpp
@@ -8,20 +8,25 @@ int main() {
 
 	// long ans = (a * b * c) - (d * e * f);
 
-	int a_new, b_new, c_new, d_new, e_new, f_new;
+	long a_new, b_new, c_new, d_new, e_new, f_new;
 
-	int div = 998244353;
-	a_new = (int)(a%(unsigned long)div);
-	b_new = (int)(b%(unsigned long)div);
-	c_new = (int)(c%(unsigned long)div);
-	d_new = (int)(d%(unsigned long)div);
-	e_new = (int)(e%(unsigned long)div);
-	f_new = (int)(f%(unsigned long)div);
+	long div = 998244353; // this is within the range of int
+	a_new = (a%(unsigned long)div);
+	b_new = (b%(unsigned long)div);
+	c_new = (c%(unsigned long)div);
+	d_new = (d%(unsigned long)div);
+	e_new = (e%(unsigned long)div);
+	f_new = (f%(unsigned long)div);
 
 	// cout << a_new << " " << b_new << " " << c_new << " " << d_new << " " << e_new << " " << f_new << endl;
 
 	// e.g., 998244352 % 998244353 will be 998244352 (but not -1), so it will cause digit overflow
-	int ans = ((a_new*b_new*c_new) - (d_new*e_new*f_new))%div;
+	// a_new, b_new, ... will be 1 ~ 998244352
+	// however, 998244352 * 998244352 will be overflowed of int; long is OK
+	long ans = ((a_new*b_new*c_new) - (d_new*e_new*f_new))%div;
+
+	// cout << (long) 998244352 * 998244352 << endl;
+	// cout << 998244352 * 998244352 << endl;
 
 
 	cout << ans << endl;
